requirements.txt
accelerate==1.12.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
decord==0.6.0
diffusers==0.36.0
easydict==1.13
exceptiongroup==1.3.1
filelock==3.20.3
fire==0.5.0
fonttools==4.61.1
fsspec==2026.1.0
gdown==5.2.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numba==0.63.1
simple-lama-inpainting==0.1.2
numpy
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
opencv-python-headless
pillow
platformdirs==4.5.1
polars==1.38.0
polars-runtime-32==1.38.0
pooch==1.9.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.15
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-image
scipy==1.15.3
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
soupsieve==2.8.3
stringzilla==4.6.0
sympy==1.14.0
termcolor==3.3.0
tifffile==2025.5.10
timm==1.0.24
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.0.0
transparent-background==1.3.4
triton==3.6.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.11
ultralytics-thop==2.0.18
urllib3==2.6.3
wget==3.2
xformers==0.0.34
zipp==3.23.0
